students/04/nvidovic/requirements.txt
python-dotenv
google-generativeai
steel-sdk
requests
beautifulsoup4